// src/components/types/enums.ts
export enum objectTypes {
  WINDOW = "window",
  WINDOW_EDGE = "window-edge",
  TOP_BAR = "top-bar"
}

export enum windowEdges {
  LEFT = "left",
  TOP = "top",
  RIGHT = "right",
  BOTTOM = "bottom",
  TOP_RIGHT = "top-right",
  TOP_LEFT = "top-left",
  BOTTOM_RIGHT = "bottom-right",
  BOTTOM_LEFT = "bottom-left",
}

export enum ActionType {
  Move,
  Resize,
}